fix: type mismatches in state machine and all test files
Fixed CI compilation errors:
- Cast config.blackout_voltage_limit from f64 to f32 for comparison
  with dcin_voltage measurements (which are f32)
- Updated all test functions across handler modules to wrap device
  and config in Arc types before passing to AppState::new()

Files updated:
- halpid/src/state_machine/machine.rs: Added f32 casts for voltage comparisons
- halpid/src/server/handlers/config.rs: Fixed 3 test functions
- halpid/src/server/handlers/usb.rs: Fixed 3 test functions
- halpid/src/server/handlers/health.rs: Fixed 1 test function
- halpid/src/server/handlers/shutdown.rs: Fixed 1 test function
- halpid/src/server/handlers/values.rs: Fixed 2 test functions

All tests now pass locally with proper type conversions.

+++ halpid/src/state_machine/machine.rs
@@ -115,7 +115,7 @@ impl StateMachine {
                 };
 
                 // Check for blackout
-                if v_in < config.blackout_voltage_limit {
+                if v_in < config.blackout_voltage_limit as f32 {
                     warn!(
                         "Detected blackout (V_in = {:.2}V < {:.2}V)",
                         v_in, config.blackout_voltage_limit
@@ -137,7 +137,7 @@ impl StateMachine {
                 };
 
                 // Check for power restoration
-                if v_in > config.blackout_voltage_limit {
+                if v_in > config.blackout_voltage_limit as f32 {
                     info!("Power resumed (V_in = {:.2}V)", v_in);
                     self.blackout_start = None;
                     self.transition_to(DaemonState::Ok);
